core: Allow Session Manager to provide new ID's
The ideal ID format might depend on the storage being used by the session
manager.

Allow it to generate ID's in the appropriate format, rather than using an
internal generator.
