+++ core/oidc.go
@@ -2,8 +2,6 @@ package core
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -35,6 +33,9 @@ type Session interface {
 // SessionManager is used to track the state of the session across it's
 // lifecycle.
 type SessionManager interface {
+	// NewID should return a new, unique identifier to be used for a session. It
+	// should be hard to guess/brute force
+	NewID() string
 	// GetSession should return the current session state for the given session
 	// ID. It should be deserialized/written in to into. If the session does not
 	// exist, found should be false with no error.
@@ -204,7 +205,7 @@ func (o *OIDC) StartAuthorization(w http.ResponseWriter, req *http.Request) (*Au
 	}
 
 	sess := &corev1beta1.Session{
-		Id:        mustGenerateID(),
+		Id:        o.smgr.NewID(),
 		Stage:     corev1beta1.Session_REQUESTED,
 		ClientId:  authreq.ClientID,
 		Request:   ar,
@@ -698,15 +699,6 @@ func (o *OIDC) Userinfo(w http.ResponseWriter, req *http.Request, handler func(w
 	}
 
 	return nil
-}
-
-// mustGenerateID returns a new, unique identifier. If it can't, it will panic
-func mustGenerateID() string {
-	b := make([]byte, 16)
-	if _, err := rand.Read(b); err != nil {
-		panic(fmt.Errorf("can't create ID, rand.Read failed: %w", err))
-	}
-	return base64.RawURLEncoding.EncodeToString(b)
 }
 
 func strsContains(strs []string, s string) bool {
